pyproject.toml
[project]
name = "gru-connect-analytics"
version = "0.1.0"
description = "Pipeline de dados para análise de conexões em Guarulhos usando Spark e dbt."
authors = [
    {name = "Paco de Oliveira Saavedra", email = "pacodeoliveirasaavedra@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "pyspark>=3.5.0",
    "dbt-spark[all]>=1.8.0",
    "apache-airflow>=2.8.0",
    "requests>=2.31.0"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"